refactor: how it works (#585)
