+++ src/components/HomePageContent/HowItWorksSection/PersonaModal.tsx
@@ -1,25 +1,35 @@
 import React, { FunctionComponent, useState } from "react";
 import { Link } from "react-router-dom";
 import { useOverlayContext, OverlayContent } from "@govtechsg/tradetrust-ui-components";
-import { PersonaProps } from "../../../types";
+import { PersonaProps } from "./types";
 import { Steps } from "./Steps";
-import { ContentType } from "../../../types";
+import { ContentType } from "./types";
 
-export const PersonaModal: FunctionComponent<PersonaProps> = ({ personaIndex, details }) => {
+const ProcessLegends = () => {
+  return (
+    <div className="hidden absolute top-0 right-0 lg:block">
+      <div className="flex justify-end">
+        <p>Manual Process</p>
+        <div className="lg:w-8 lg:mt-3 lg:ml-3.5 lg:border-t lg:border-white lg:border-dashed" />
+      </div>
+      <div className="flex justify-end">
+        <p>Digital Process</p>
+        <div className="lg:w-8 lg:mt-3 lg:ml-3.5 lg:border-t lg:border-white lg:border-solid" />
+      </div>
+    </div>
+  );
+};
+
+export const PersonaModal: FunctionComponent<PersonaProps> = ({ personaIndex, persona }) => {
   const { setOverlayVisible, showOverlay } = useOverlayContext();
   const [selectedContentType, setSelectedContentType] = useState<ContentType>(ContentType.THEN);
-  const contentType: ContentType[] = [ContentType.THEN, ContentType.NOW];
+  const contentType: ContentType[] = [ContentType.THEN, ContentType.NOW]; // ContentType.BENEFIT omitted here, but manually rendered later?
   const handleCloseOverlay = (): void => {
     setOverlayVisible(false);
     showOverlay(undefined);
   };
 
-  const { title, thenSteps, nowSteps, startMessage, benefits, endMessage } = details.learnMore;
-
-  const contentTypeFilterStyle = (item: ContentType): string => {
-    const returnStyle = item === ContentType.THEN ? "mr-5" : "";
-    return item === selectedContentType ? `${returnStyle} font-gilroy-bold underline cursor-default` : returnStyle;
-  };
+  const { title, thenSteps, nowSteps, startMessage, benefits, endMessage } = persona.learnMore;
 
   return (
     <section id="persona-modal">
@@ -31,44 +41,40 @@ export const PersonaModal: FunctionComponent<PersonaProps> = ({ personaIndex, de
           <div className="flex flex-col justify-center">
             <div className="relative flex justify-center w-full">
               <h3 className="text-center">{title}</h3>
-              {thenSteps && nowSteps && (
-                <div className="hidden absolute bottom-0 right-0 lg:block">
-                  <div className="flex justify-end">
-                    <p>Manual Process</p>
-                    <div className="lg:w-8 lg:mt-3 lg:ml-3.5 lg:border-t lg:border-white lg:border-dashed" />
-                  </div>
-                  <div className="flex justify-end">
-                    <p>Digital Process</p>
-                    <div className="lg:w-8 lg:mt-3 lg:ml-3.5 lg:border-t lg:border-white lg:border-solid" />
-                  </div>
-                </div>
-              )}
             </div>
             {startMessage && <h4 className="text-center mt-8">{startMessage}</h4>}
             {thenSteps && nowSteps && (
-              <div className="flex flex-row justify-center lg:hidden">
-                {contentType.map((content, index) => (
-                  <h3
-                    key={index}
-                    className={`font-normal text-lemon-500 text-center mt-12 cursor-pointer ${contentTypeFilterStyle(
-                      content
-                    )}`}
-                    onClick={() => setSelectedContentType(content)}
-                  >
-                    {content}
-                  </h3>
-                ))}
-              </div>
-            )}
-            {thenSteps &&
-              nowSteps &&
-              contentType.map((content, index) => (
-                <div key={index} className="flex flex-col justify-center">
-                  <div className={`lg:inline ${content === selectedContentType ? "inline" : "hidden"}`}>
-                    <Steps contentType={content} stepsDetails={content === ContentType.THEN ? thenSteps : nowSteps} />
-                  </div>
+              <>
+                <ProcessLegends />
+                <div className="flex flex-row justify-center lg:hidden">
+                  {contentType.map((content, index) => {
+                    const cssState = content === selectedContentType ? "font-gilroy-bold underline" : "font-normal";
+                    const cssAlign = content === ContentType.THEN ? "mr-5" : "";
+                    return (
+                      <h3
+                        key={index}
+                        className={`text-lemon-500 text-center mt-12 cursor-pointer ${cssState} ${cssAlign}`}
+                        onClick={() => setSelectedContentType(content)}
+                      >
+                        {content}
+                      </h3>
+                    );
+                  })}
                 </div>
-              ))}
+                <>
+                  {contentType.map((content, index) => (
+                    <div key={index} className="flex flex-col justify-center">
+                      <div className={`lg:inline ${content === selectedContentType ? "inline" : "hidden"}`}>
+                        <Steps
+                          contentType={content}
+                          stepsDetails={content === ContentType.THEN ? thenSteps : nowSteps}
+                        />
+                      </div>
+                    </div>
+                  ))}
+                </>
+              </>
+            )}
             {benefits && <Steps contentType={ContentType.BENEFIT} stepsDetails={benefits} />}
             <h4 className="text-center mt-8">{endMessage}</h4>
           </div>
